requirements.txt
dash==2.10.2
dash_iconify==0.1.2
exifread==3.0.0
feffery_antd_components==0.3.10
feffery_leaflet_components==0.1.0rc8
feffery_utils_components==0.1.29
Flask==2.0.2
geopy==2.4.1
oss2==2.19.0
Pillow==11.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
reverse_geocoder==1.5.1
watchdog==5.0.3
gunicorn==23.0.0
scipy==1.9.3